DNM: ARM64/dts: qcom: msm8953-thermal: Increase cpu trip temperature
We have seriously issue with mitigation on frequency when already reach the limit, it seems
cpus_trip was taken part of this mitigation that always do throttle when case-therm-step sensor
already reach 43 in (millicelcius).

For workaround, i try to increase that cpu trip limit at 60 in (millicelcius). And see if that can
bypass that very low temperature mitigation limit, so it won't lock freq at 1.6Ghz anymore for some
times.

Another approach is to change bind freq from nodes "cooling-device = <&CPU0 THERMAL_NO_LIMIT 3>" on
cooling-maps in case-therm-step. "3" Stand as frequency step that qualcomm choose to limit at 1.6
/* throttle from fmax to 1689600KHz */, if we can configure this limit at suitable limit or atleast
remove it in dirty hacks. This can be remove that thermal limitation with several impact that may happen.

For now increase cpu trip temperature on testing and let's see if we have this issue or not after get
a long deepsleep issue.

References: 6cabe9fca1d918776c830961e3314674880f2b43
Signed-off-by: Dicky Herlambang <herlambangdicky5@gmail.com>